main: Sanitize safe-mode shutdown
Safe-mode placeholder process is restarted every time it dies - but that
is not good if init is shutting down the system. This process should not
be terminated during shutdown - in fact, if a 'safe-shutdown' process
crashes, safe-mode should be called. This patch prevents init from
terminating safe-mode process - be it just the placeholder or the real
one.

An interesting question is: should safe-mode process be somehow informed
that system is going to shutdown? This is not addressed in this patch.
It will simply die on reboot system call, right now.

+++ src/main.c
@@ -777,9 +777,11 @@ static enum timeout_result kill_timeout_cb(void)
 		log_message("Sending KILL signal to processes that refused to "
 			    "term in timeout\n");
 		while (p != NULL) {
-			log_message("Sending KILL signal to %d (%s)\n", p->pid,
-				    p->config.process_name);
-			kill(p->pid, SIGKILL);
+			if (p->config.type != SAFE_MODE) {
+				log_message("Sending KILL signal to %d (%s)\n",
+					    p->pid, p->config.process_name);
+				kill(p->pid, SIGKILL);
+			}
 			p = p->next;
 		}
 	}
@@ -792,9 +794,11 @@ static void term_running_process(void)
 {
 	struct process *p = running_processes;
 	while (p != NULL) {
-		log_message("Sending TERM signal to %d (%s)\n", p->pid,
-			    p->config.process_name);
-		kill(p->pid, SIGTERM);
+		if (p->config.type != SAFE_MODE) {
+			log_message("Sending TERM signal to %d (%s)\n", p->pid,
+				    p->config.process_name);
+			kill(p->pid, SIGTERM);
+		}
 		p = p->next;
 	}
